Step size is now applied before the projection of the Jt gradient to the tangent plane of the constraint function h and not after.
The transient behavior is quite unsteady although I initiated p_ref such that there is power balance.
The problem is that when I start with all voltages v=1, then the first retraction step increases v to more than 6 and theta stays nearly zero.
But when I initialise v with 0.1 then the voltage stays low in the first step but theta changes more in order to achieve the desired power flow.

Signed-off-by: Felix Boewing <boewingf@student.ethz.ch>